fix: primitive 타입 적용

+++ src/main/java/com/palangwi/soup/dto/keyword/response/SearchKeywordsResponseDto.java
@@ -4,11 +4,11 @@ import java.util.List;
 
 public record SearchKeywordsResponseDto(
         List<SearchKeywordDto> keywords,
-        int totalElements,
+        long totalElements,
         int totalPages,
         int currentPages) {
 
-    public static SearchKeywordsResponseDto from(List<SearchKeywordDto> keywords, int totalElements, int totalPages,
+    public static SearchKeywordsResponseDto from(List<SearchKeywordDto> keywords, long totalElements, int totalPages,
             int currentPages) {
         return new SearchKeywordsResponseDto(keywords, totalElements, totalPages, currentPages);
     }
